rename mapReducer to positionReducer
