+++ src/services/StoreService.js
@@ -1,5 +1,5 @@
 import { combineReducers, createStore } from 'redux';
-import { mapReducer, initialState as initialMapState, ZOOM_CHANGED, POSITION_CHANGED } from '../modules/map/store/olMap.reducer';
+import { positionReducer, initialState as initialMapState, ZOOM_CHANGED, POSITION_CHANGED } from '../modules/map/store/position.reducer';
 import { sidePanelReducer } from '../modules/menue/store/sidePanel.reducer';
 import { modalReducer } from '../modules/modal/store/modal.reducer';
 import { contextMenueReducer } from '../modules/contextMenue/store/contextMenue.reducer';
@@ -22,7 +22,7 @@ export class StoreService {
 		const storeEnhancer = ReduxQuerySync.enhancer({
 			params: {
 				zoom: {
-					selector: state => state.map.zoom,
+					selector: state => state.position.zoom,
 
 					action: value => ({ type: ZOOM_CHANGED, payload: value }),
 
@@ -42,7 +42,7 @@ export class StoreService {
 					defaultValue: initialMapState.zoom,
 				},
 				position: {
-					selector: state => state.map.position,
+					selector: state => state.position.position,
 					action: value => ({ type: POSITION_CHANGED, payload: value }),
 
 					//TODO: handler non parseable string
@@ -66,7 +66,7 @@ export class StoreService {
 			 * must be named like the field of the state
 			 * see: https://redux.js.org/recipes/structuring-reducers/initializing-state#combined-reducers
 			 */
-			map: mapReducer,
+			position: positionReducer,
 			sidePanel: sidePanelReducer,
 			contextMenue: contextMenueReducer,
 			modal:modalReducer,
