+++ test/modules/map/components/infoButton/InfoButton.test.js
@@ -1,11 +1,11 @@
 /* eslint-disable no-undef */
 
 import { InfoButton } from '../../../../../src/modules/map/components/infoButton/InfoButton';
-import { mapReducer } from '../../../../../src/modules/map/store/olMap.reducer';
+import { positionReducer } from '../../../../../src/modules/map/store/position.reducer';
 import { TestUtils } from '../../../../test-utils.js';
 import { $injector } from '../../../../../src/injection';
 import { Popup } from '../../../../../src/modules/commons/components/popup/Popup';
-import { changePosition, changeZoom } from '../../../../../src/modules/map/store/olMap.action';
+import { changePosition, changeZoom } from '../../../../../src/modules/map/store/position.action';
 
 window.customElements.define(InfoButton.tag, InfoButton);
 window.customElements.define(Popup.tag, Popup);
@@ -17,13 +17,13 @@ describe('InfoButton', () => {
 	beforeEach(async () => {
 
 		state = {
-			map: {
+			position: {
 				zoom: 5,
 				position: [1288239.2412306187, 6130212.561641981]
 			}
 		};
 
-		store = TestUtils.setupStoreAndDi(state, { map: mapReducer });
+		store = TestUtils.setupStoreAndDi(state, { position: positionReducer });
 		$injector
 			.registerSingleton('TranslationService', { translate: (key) => key });
 
@@ -92,7 +92,7 @@ describe('InfoButton', () => {
 			expect(element.shadowRoot.getElementById('info-popup').getAttribute('type')).toBe('show');
 			expect(element.shadowRoot.getElementById('info-popup').isOpen()).toBeTrue();
 
-			expect(store.getState().map.zoom).toBe(5); 
+			expect(store.getState().position.zoom).toBe(5); 
 
 			changeZoom(4);
 
@@ -112,7 +112,7 @@ describe('InfoButton', () => {
 			expect(element.shadowRoot.getElementById('info-popup').getAttribute('type')).toBe('show');
 			expect(element.shadowRoot.getElementById('info-popup').isOpen()).toBeTrue();
 
-			expect(store.getState().map.position).toEqual([1288239.2412306187, 6130212.561641981]); 
+			expect(store.getState().position.position).toEqual([1288239.2412306187, 6130212.561641981]); 
 
 			changePosition([1290570.5705933168, 6129218.880274274]);
