+++ test/modules/utils/components/showCase/ShowCase.test.js
@@ -2,7 +2,7 @@ import { ShowCase } from '../../../../../src/modules/utils/components/showCase/S
 import { Toggle } from '../../../../../src/modules/commons/components/toggle/Toggle';
 import { TestUtils } from '../../../../test-utils';
 import { sidePanelReducer } from '../../../../../src/modules/menue/store/sidePanel.reducer';
-import { mapReducer } from '../../../../../src/modules/map/store/olMap.reducer';
+import { positionReducer } from '../../../../../src/modules/map/store/position.reducer';
 import { layersReducer } from '../../../../../src/modules/map/store/layers/layers.reducer';
 import { OlCoordinateService } from '../../../../../src/services/OlCoordinateService';
 import { $injector } from '../../../../../src/injection';
@@ -17,7 +17,7 @@ describe('ShowCase', () => {
 	const setup = () => {
 		const state = {
 			mobile: false,
-			map: {
+			position: {
 				zoom: 10,
 				pointerPosition: [1288239.2412306187, 6130212.561641981]
 			},
@@ -29,7 +29,7 @@ describe('ShowCase', () => {
 				background:'null'
 			}
 		};
-		store = TestUtils.setupStoreAndDi(state, { map: mapReducer, sidePanel: sidePanelReducer, layers:layersReducer });
+		store = TestUtils.setupStoreAndDi(state, { position: positionReducer, sidePanel: sidePanelReducer, layers:layersReducer });
 		$injector
 			.register('CoordinateService', OlCoordinateService)
 			.registerSingleton('EnvironmentService', { isEmbedded : () => false });
@@ -64,7 +64,7 @@ describe('ShowCase', () => {
 			
 			element.shadowRoot.querySelector('#button0').click();
 			
-			expect(store.getState().map.zoom).toBe(13);
+			expect(store.getState().position.zoom).toBe(13);
 		});
 
 		it('calls the callback, if button1 are clicked', async () => {
@@ -72,7 +72,7 @@ describe('ShowCase', () => {
 			
 			element.shadowRoot.querySelector('#button1').click();
 			
-			expect(store.getState().map.zoom).toBe(11);
+			expect(store.getState().position.zoom).toBe(11);
 		});
 
 		it('toggle gets checked, if toggle are clicked', async () => {
